fix realloc_com_reqs
So far alloc_com was always called as:
  com = alloc_com(sizeof(double), 1);
so that com->entrysize was sizeof(double)=8.

We now do not use com->entrysize for anything any more, because we have
replaced stuff like
  com->send_buf = realloc(com->send_buf, n_rq_new*com->entrysize);
by
  com->send_buf = realloc(com->send_buf, n_rq_new*sizeof(void *));

Since sizeof(void *) = sizeof(double) = 8 this changes nothing, but will
keep working if the first argument of alloc_com is no longer 8.