compute tags in surface.c and indicators.c differently
we had:
       s_ltag = (nb_lid*64 + nb_ni)*6 + nb_f;
       r_ltag = (lid*64 + ni)*6 + f;
now we just use:
       s_ltag = nb_lid*6 + nb_f;
       r_ltag = lid*6 + f;
in request_surfaces_exchange_for_all_vars and request_indc_exchange_for_vl.

We do this because ni and nb_ni do not necessarily agree on different ranks,
as the elms in mesh->nbelm have only the fnb needed for communication and
not all of them.