possible fix for Persson trouble indicator
Until now evolve_Persson_trouble_ncoeffs_dg used all fv points for
interplation. This introduced numerical noise that the trouble score picks
up as shocks.
We now use a max of 12th order Lagrange interp if we are on a fv node. This
way the numerical noise should be gone. And then it is more likely to switch
to dg in case all is smooth.