use more points in find_elmfacepoints_in_nbface
Before we used 3*3 points on a face, that were not close to the edges.
Now we use 6*6 points and put the outermost ones just 1e-5 from the edges.
This way it can better find if a face touches a nb face.

Note: If find_elmfacepoints_in_nbface fails to detect that two faces
actually touch the nb-info becomes wrong. When we then refine some ranks may
be able to update their nb-info, while other may not be able to do so. This
results in inconsistent nb-info between MPI ranks. Then we get MPI errors like
"buffer trnucated".